requirements.txt
Flask
torch
torchvision
torchaudio
diffusers[torch]
transformers
accelerate
pillow
safetensors
huggingface-hub
numpy
gunicorn
